backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart>=0.0.5
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
nltk>=3.6.0
textstat>=0.7.3
joblib>=1.0.0
python-dotenv>=0.19.0 